Base the local html on index.html
This way style and other bits added to index.html will be reflected both when running locally and with pyscript. But currently it assumes a specific template for the body.

I needed to make the <meta /> and <link /> tags in index.html self closing for the minidom parser to work. Unfortunately, this style will have to be followed in all future versions of index for the parser to keep working.